Split state block iter into its own module
